backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
bcrypt
sqlalchemy
python-multipart
matplotlib
numpy
